+++ src/distances/cv.hpp
@@ -19,9 +19,10 @@ std::vector<std::string> all_contexts{};
 
 template <seqan3::alphabet alphabet_t>
 std::vector<std::string>
-get_all_contexts(int order, robin_hood::unordered_set<int> &valid_characters) {
-  int n_valid_characters = valid_characters.size();
-  int n_kmers = std::pow(n_valid_characters, order);
+get_all_contexts(size_t order,
+                 robin_hood::unordered_set<size_t> &valid_characters) {
+  size_t n_valid_characters = valid_characters.size();
+  size_t n_kmers = std::pow(n_valid_characters, order);
 
   if (all_contexts.size() == n_kmers) {
     return all_contexts;
@@ -29,16 +30,17 @@ get_all_contexts(int order, robin_hood::unordered_set<int> &valid_characters) {
 
   std::vector<std::string> contexts(n_kmers, std::string(order, ' '));
 
-  for (int n = 0; n < order; n++) {
-    int n_char = 0;
+  for (size_t n = 0; n < order; n++) {
+    size_t n_char = 0;
     for (auto &char_rank : valid_characters) {
-      int step_size = n_kmers / int(std::pow(n_valid_characters, (n + 1)));
+      size_t step_size =
+          n_kmers / size_t(std::pow(n_valid_characters, (n + 1)));
 
-      int start = step_size * n_char;
-      int stop = step_size * (n_char + 1);
+      size_t start = step_size * n_char;
+      size_t stop = step_size * (n_char + 1);
 
       while (stop <= n_kmers) {
-        for (int i = start; i < stop; i++) {
+        for (size_t i = start; i < stop; i++) {
           alphabet_t c = seqan3::assign_rank_to(char_rank, alphabet_t{});
           contexts[i][n] = c.to_char();
         }
@@ -102,7 +104,8 @@ float cosine_dissimilarity(Eigen::VectorXd &left, Eigen::VectorXd &right) {
 template <seqan3::alphabet alphabet_t>
 inline float core_cv(ProbabilisticSuffixTreeMap<alphabet_t> &left,
                      ProbabilisticSuffixTreeMap<alphabet_t> &right,
-                     std::vector<std::string> &contexts, int background_order) {
+                     std::vector<std::string> &contexts,
+                     size_t background_order) {
   static std::unordered_map<std::string, Eigen::VectorXd> cache{};
 
   Eigen::VectorXd left_composition_vector =
@@ -120,7 +123,7 @@ namespace pst::distances {
 template <seqan3::alphabet alphabet_t>
 inline float cv(ProbabilisticSuffixTreeMap<alphabet_t> &left,
                 ProbabilisticSuffixTreeMap<alphabet_t> &right,
-                int background_order = 2) {
+                size_t background_order = 2) {
   auto contexts = details::get_shared_contexts(left, right);
   return details::core_cv<alphabet_t>(left, right, contexts, background_order);
 }
@@ -128,7 +131,7 @@ inline float cv(ProbabilisticSuffixTreeMap<alphabet_t> &left,
 template <seqan3::alphabet alphabet_t>
 inline float cv_estimation(ProbabilisticSuffixTreeMap<alphabet_t> &left,
                            ProbabilisticSuffixTreeMap<alphabet_t> &right,
-                           int order = 6, int background_order = 2) {
+                           size_t order = 6, size_t background_order = 2) {
   if (details::all_contexts.empty()) {
     details::get_all_contexts<alphabet_t>(order, left.valid_characters);
   }
