+++ src/distances/score.hpp
@@ -18,13 +18,13 @@ namespace pst {
 using tree_t = pst::ProbabilisticSuffixTreeMap<seqan3::dna5>;
 
 void score_slices(
-    int start_index, int stop_index, std::vector<std::vector<double>> &scores,
-    std::vector<tree_t> &trees,
+    size_t start_index, size_t stop_index,
+    std::vector<std::vector<double>> &scores, std::vector<tree_t> &trees,
     std::vector<std::vector<seqan3::dna5>> &sequences,
     const std::function<float(tree_t &, std::vector<seqan3::dna5> &)> &fun) {
 
-  for (int j = 0; j < sequences.size(); j++) {
-    for (int i = start_index; i < stop_index; i++) {
+  for (size_t j = 0; j < sequences.size(); j++) {
+    for (size_t i = start_index; i < stop_index; i++) {
       scores[j][i] = fun(trees[i], sequences[j]);
     }
   }
@@ -32,19 +32,20 @@ void score_slices(
 
 std::vector<std::vector<double>>
 score_sequences(std::vector<tree_t> &trees, std::vector<std::string> &sequences,
-                int background_order) {
+                size_t background_order) {
   std::vector<std::vector<double>> scores(sequences.size(),
                                           std::vector<double>(trees.size()));
 
   std::vector<std::vector<seqan3::dna5>> dna_sequences(sequences.size());
-  for (int sequence_idx = 0; sequence_idx < sequences.size(); sequence_idx++) {
+  for (size_t sequence_idx = 0; sequence_idx < sequences.size();
+       sequence_idx++) {
     std::vector<seqan3::dna5> seq = sequences[sequence_idx] |
                                     seqan3::views::char_to<seqan3::dna5> |
                                     seqan3::views::to<std::vector>;
     dna_sequences[sequence_idx] = seq;
   }
 
-  auto fun = [&](int start_index, int stop_index) {
+  auto fun = [&](size_t start_index, size_t stop_index) {
     score_slices(
         start_index, stop_index, std::ref(scores), std::ref(trees),
         std::ref(dna_sequences),
